requirements.txt
autopep8==1.5.3
beautifulsoup4==4.9.1
selenium==3.141.0
python-dotenv==0.10.5
Flask==1.1.1
Flask-HTTPAuth==3.3.0